Corrected wrong calculation of pldist()
Source: https://github.com/fhirschmann/rdp/issues/13.

The current pldist function returns the distance between a point and the infinte line generated by two other points (i.e the distance to the orthogonal projection), which is incorrect. It should return the distance between a point and a line-segment.
Here is a fix coming from this answer https://stackoverflow.com/a/54442561.